A2/Part B/microservice1/requirements.txt
Flask==2.0.2
requests==2.26.0
google-cloud-firestore==2.1.0
Flask-Cors==3.0.10